frontend/latex/output: more refactor and fix content→contents, note when there is no content in latex, and actually update it (wrong frame type hits again)
