+++ src/packages/frontend/frame-editors/latex-editor/actions.ts
@@ -1850,7 +1850,7 @@ export class Actions extends BaseActions<LatexEditorState> {
     if (
       !force &&
       !this.get_matching_frame({ type: "latex_table_of_contents" }) &&
-      !this.get_matching_frame({ type: "latex-output" })
+      !this.get_matching_frame({ type: "output" })
     ) {
       // There is no table of contents frame or output frame so don't update that info.
       return;
