+++ src/packages/frontend/frame-editors/latex-editor/output.tsx
@@ -21,7 +21,7 @@ With build controls at the top (build, force build, clean, etc.)
 import type { Data } from "@cocalc/frontend/frame-editors/frame-tree/pinch-to-zoom";
 import type { TabsProps } from "antd";
 
-import { List as AntdList, Avatar, Button, Spin, Tabs, Tag } from "antd";
+import { Alert, Space, Spin, Tabs, Tag } from "antd";
 import { List } from "immutable";
 import { useCallback, useMemo, useState } from "react";
 import { defineMessage, useIntl } from "react-intl";
@@ -31,14 +31,11 @@ import {
   Icon,
   TableOfContents,
   TableOfContentsEntryList,
+  Text,
 } from "@cocalc/frontend/components";
-import StaticMarkdown from "@cocalc/frontend/editors/slate/static-markdown";
-import { filenameIcon } from "@cocalc/frontend/file-associations";
 import { EditorState } from "@cocalc/frontend/frame-editors/frame-tree/types";
 import { project_api } from "@cocalc/frontend/frame-editors/generic/client";
-import { editor, labels } from "@cocalc/frontend/i18n";
-import { path_split, plural } from "@cocalc/util/misc";
-import { COLORS } from "@cocalc/util/theme";
+import { editor } from "@cocalc/frontend/i18n";
 
 import { Actions } from "./actions";
 import { Build } from "./build";
@@ -46,9 +43,11 @@ import { WORD_COUNT_ICON } from "./constants";
 import { ErrorsAndWarnings } from "./errors-and-warnings";
 import { use_build_logs } from "./hooks";
 import { PDFControls } from "./output-control";
+import { OutputFiles } from "./output-files";
+import { OutputStats } from "./output-stats";
 import { PDFJS } from "./pdfjs";
-import { useTexSummaries } from "./use-summarize";
 import { BuildLogs } from "./types";
+import { useTexSummaries } from "./use-summarize";
 
 interface OutputProps {
   id: string;
@@ -67,24 +66,11 @@ interface OutputProps {
 
 type TabType = "pdf" | "contents" | "files" | "build" | "errors" | "stats";
 
-interface FileListItem {
-  path: string;
-  displayPath: string;
-  isMain: boolean;
-  summary: string;
-}
-
 const STATS_LABEL = defineMessage({
   id: "latex.output.stats_tab.label",
   defaultMessage: "Stats",
   description:
-    "Short abbreviation for 'Statistics' used as tab label. Should be abbreviated like 'Stats' in English for 'Statistics', 'Stats' in German, or 'Estad' in Spanish - a recognizable short form, not the full word.",
-});
-
-const STATISTICS_HEADER = defineMessage({
-  id: "latex.output.stats.header",
-  defaultMessage: "Statistics",
-  description: "Header text for the statistics section in LaTeX output",
+    "Short abbreviation for 'Statistics' used as tab label. Should be abbreviated like 'Stats' in English for 'Stats' in German, or 'Estad' in Spanish - a recognizable short form, not the full word.",
 });
 
 export function Output(props: OutputProps) {
@@ -213,6 +199,13 @@ export function Output(props: OutputProps) {
     }
   }, [activeTab, reload, refreshWordCount]);
 
+  // Refresh TOC when contents tab is opened
+  useEffect(() => {
+    if (activeTab === "contents") {
+      actions.updateTableOfContents(true);
+    }
+  }, [activeTab, actions]);
+
   // Sync state with stored values when they change
   React.useEffect(() => {
     setActiveTab(storedTab);
@@ -306,7 +299,6 @@ export function Output(props: OutputProps) {
   }, [build_logs, knitr]);
 
   // No automatic tab switching - let user control tabs manually
-  // Errors are indicated with red exclamation icon only
 
   function renderPdfTab() {
     return {
@@ -394,142 +386,21 @@ export function Output(props: OutputProps) {
             contents={contents}
             fontSize={uiFontSize}
             scrollTo={actions.scrollToHeading.bind(actions)}
+            ifEmpty={
+              <Alert
+                type="info"
+                message="Table of Contents is empty"
+                description={
+                  <>
+                    Add <Text code>{"\\section{...}"}</Text> and{" "}
+                    <Text code>{"\\subsection{...}"}</Text> commands to your
+                    LaTeX document to create a table of contents.
+                  </>
+                }
+                style={{ margin: "15px" }}
+              />
+            }
           />
-        </div>
-      ),
-    };
-  }
-
-  function renderFilesTab() {
-    // Filter out the main file from the list
-    const subFiles = switch_to_files
-      .filter((filePath) => filePath !== path)
-      .sort();
-    const subFileCount = subFiles.size;
-
-    // Compute the common prefix to strip (directory of main file)
-    const prefix = path_split(path).head;
-    const prefixWithSlash = prefix ? prefix + "/" : "";
-
-    const listData = subFiles.toJS().map((filePath: string) => {
-      const displayPath = filePath.startsWith(prefixWithSlash)
-        ? filePath.slice(prefixWithSlash.length)
-        : filePath;
-      return {
-        path: filePath,
-        displayPath,
-        isMain: false,
-        summary: fileSummaries[filePath] ?? "Loading...",
-      };
-    });
-
-    return {
-      key: "files",
-      label: (
-        <span style={{ display: "flex", alignItems: "center", gap: "2px" }}>
-          {summariesLoading ? <Spin size="small" /> : <Icon name="file" />}
-          Files
-        </span>
-      ),
-      children: (
-        <div
-          className="smc-vfill"
-          style={{
-            display: "flex",
-            flexDirection: "column",
-            height: "100%",
-          }}
-        >
-          {/* Fixed header with buttons and file count */}
-          <div
-            style={{
-              display: "flex",
-              justifyContent: "space-between",
-              alignItems: "center",
-              padding: "10px",
-              borderBottom: "1px solid #d9d9d9",
-              backgroundColor: "white",
-              flexShrink: 0,
-            }}
-          >
-            <Button
-              type="primary"
-              size="small"
-              icon={<Icon name="tex-file" />}
-              onClick={() => actions.switch_to_file(path)}
-            >
-              Open Main File
-            </Button>
-
-            <span style={{ color: COLORS.GRAY_M, fontSize: uiFontSize }}>
-              {subFileCount} {plural(subFileCount, "subfile")}
-            </span>
-
-            <Button
-              size="small"
-              icon={<Icon name="refresh" />}
-              onClick={refreshSummaries}
-              loading={summariesLoading}
-              disabled={summariesLoading}
-            >
-              {intl.formatMessage(labels.refresh)}
-            </Button>
-          </div>
-
-          {/* Scrollable list */}
-          <div
-            style={{
-              flex: 1,
-              overflowY: "auto",
-              padding: "10px",
-            }}
-          >
-            <AntdList
-              size="small"
-              dataSource={listData}
-              renderItem={(item: FileListItem) => (
-                <AntdList.Item
-                  style={{
-                    cursor: "pointer",
-                  }}
-                  onClick={() => actions.switch_to_file(item.path)}
-                >
-                  <AntdList.Item.Meta
-                    avatar={
-                      <Avatar
-                        size="default"
-                        style={{
-                          backgroundColor: "transparent",
-                          color: COLORS.GRAY_D,
-                        }}
-                        icon={<Icon name={filenameIcon(item.path)} />}
-                      />
-                    }
-                    title={
-                      <span
-                        style={{
-                          fontFamily: "monospace",
-                          fontSize: `${uiFontSize}px`,
-                        }}
-                      >
-                        {item.displayPath}
-                      </span>
-                    }
-                    description={
-                      <span
-                        style={{
-                          color: COLORS.GRAY_M,
-                          fontSize: uiFontSize - 2,
-                        }}
-                      >
-                        <StaticMarkdown value={item.summary} />
-                      </span>
-                    }
-                  />
-                </AntdList.Item>
-              )}
-            />
-          </div>
         </div>
       ),
     };
@@ -558,91 +429,7 @@ export function Output(props: OutputProps) {
     };
   }
 
-  function renderWordCountTab() {
-    return {
-      key: "stats",
-      label: (
-        <span style={{ display: "flex", alignItems: "center", gap: "2px" }}>
-          {wordCountLoading ? (
-            <Spin size="small" />
-          ) : (
-            <Icon name={WORD_COUNT_ICON} />
-          )}
-          {intl.formatMessage(STATS_LABEL)}
-        </span>
-      ),
-      children: (
-        <div
-          className="smc-vfill"
-          style={{
-            display: "flex",
-            flexDirection: "column",
-            height: "100%",
-          }}
-        >
-          {/* Fixed header with refresh button */}
-          <div
-            style={{
-              display: "flex",
-              justifyContent: "space-between",
-              alignItems: "center",
-              padding: "10px",
-              borderBottom: "1px solid #d9d9d9",
-              backgroundColor: "white",
-              flexShrink: 0,
-            }}
-          >
-            <span
-              style={{
-                color: COLORS.GRAY_M,
-                fontSize: uiFontSize,
-                display: "flex",
-                alignItems: "center",
-                gap: "4px",
-              }}
-            >
-              <Icon name={WORD_COUNT_ICON} />
-              {intl.formatMessage(STATISTICS_HEADER)}
-            </span>
-
-            <Button
-              size="small"
-              icon={<Icon name="refresh" />}
-              onClick={() => refreshWordCount(true)}
-              loading={wordCountLoading}
-              disabled={wordCountLoading}
-            >
-              {intl.formatMessage(labels.refresh)}
-            </Button>
-          </div>
-
-          {/* Scrollable statistics content */}
-          <div
-            style={{
-              flex: 1,
-              overflowY: "auto",
-              padding: "10px",
-            }}
-          >
-            <pre
-              style={{
-                fontSize: `${uiFontSize - 2}px`,
-                fontFamily: "monospace",
-                whiteSpace: "pre-wrap",
-                wordWrap: "break-word",
-                margin: 0,
-                color: COLORS.GRAY_D,
-              }}
-            >
-              {wordCount ||
-                "Click refresh to generate word count statistics..."}
-            </pre>
-          </div>
-        </div>
-      ),
-    };
-  }
-
+  // Errors are indicated with red icon only
   function renderErrorsTab() {
     const { errors, warnings, typesetting } = errorCounts;
     const hasAnyIssues = errors > 0 || warnings > 0 || typesetting > 0;
@@ -654,11 +441,11 @@ export function Output(props: OutputProps) {
           <Icon name="bug" />
           {intl.formatMessage(editor.errors_and_warnings_title_short)}
           {hasAnyIssues && (
-            <span style={{ display: "flex", gap: "2px" }}>
+            <Space.Compact>
               {errors > 0 && <Tag color="red">{errors}</Tag>}
               {warnings > 0 && <Tag color="orange">{warnings}</Tag>}
               {typesetting > 0 && <Tag color="blue">{typesetting}</Tag>}
-            </span>
+            </Space.Compact>
           )}
         </span>
       ),
@@ -680,6 +467,53 @@ export function Output(props: OutputProps) {
     };
   }
 
+  function renderFilesTab() {
+    return {
+      key: "files",
+      label: (
+        <span style={{ display: "flex", alignItems: "center", gap: "2px" }}>
+          {summariesLoading ? <Spin size="small" /> : <Icon name="file" />}
+          Files
+        </span>
+      ),
+      children: (
+        <OutputFiles
+          switch_to_files={switch_to_files}
+          path={path}
+          fileSummaries={fileSummaries}
+          summariesLoading={summariesLoading}
+          refreshSummaries={refreshSummaries}
+          actions={actions}
+          uiFontSize={uiFontSize}
+        />
+      ),
+    };
+  }
+
+  function renderStatsTab() {
+    return {
+      key: "stats",
+      label: (
+        <span style={{ display: "flex", alignItems: "center", gap: "2px" }}>
+          {wordCountLoading ? (
+            <Spin size="small" />
+          ) : (
+            <Icon name={WORD_COUNT_ICON} />
+          )}
+          {intl.formatMessage(STATS_LABEL)}
+        </span>
+      ),
+      children: (
+        <OutputStats
+          wordCountLoading={wordCountLoading}
+          wordCount={wordCount}
+          refreshWordCount={refreshWordCount}
+          uiFontSize={uiFontSize}
+        />
+      ),
+    };
+  }
+
   function renderTabs() {
     const tabItems: NonNullable<TabsProps["items"]> = [
       renderPdfTab(),
@@ -687,7 +521,7 @@ export function Output(props: OutputProps) {
       ...(switch_to_files?.size > 1 ? [renderFilesTab()] : []),
       renderBuildTab(),
       renderErrorsTab(),
-      renderWordCountTab(),
+      renderStatsTab(),
     ];
 
     return (
