+++ src/packages/frontend/i18n/trans/ar_EG.json
@@ -597,7 +597,7 @@
   "editor.snippets.label": "المقتطفات",
   "editor.snippets.tooltip": "افتح لوحة تحتوي على مقتطفات الكود",
   "editor.table_of_contents.name": "جدول المحتويات",
-  "editor.table_of_contents.short": "محتوى",
+  "editor.table_of_contents.short": "المحتويات",
   "editor.terminal.cmd.help.title": "عرض الوثائق لاستخدام محطة لينكس في CoCalc",
   "file_actions.compress.name": "ضغط",
   "file_actions.copy.name": "نسخ",
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "التركية",
   "jupyter.about.features": "<ul> <li> <b>المزامنة الفورية:</b> يمكن لعدة أشخاص تحرير الدفاتر في نفس الوقت: مؤشرات متعددة، تراجع وإعادة خاصة بكل مستخدم عبر المستند، مزامنة فورية لـ ipywidgets </li> <li> <b>تقسيم المستند:</b> تحرير ورؤية أجزاء متعددة من دفتر كبير في نفس الوقت </li> <li> <b>TimeTravel:</b> رؤية تاريخ تفصيلي لكيفية إنشاء دفتر ملاحظات بالضبط </li> <li> <b>النص/Markdown:</b> محرر رسومي قوي لخلايا markdown، بحيث يمكنك كتابة نص منسق بشكل جيد لشرح الكود الخاص بك دون الحاجة إلى كتابة markdown مباشرة. </li> <li> <b>مساعد الذكاء الاصطناعي:</b> عدة أدوات ذكاء اصطناعي تساعدك على فهم رسائل الخطأ، تحسين الكود، توليد خلايا كود جديدة بناءً على السياق، أو فقط شرح/توثيق الكود الموجود. </li> <li> <b>المقتطفات:</b> عينات كود للعديد من النوى </li> <li> <b>السبورة:</b> إنشاء سبورة (مستند ينتهي بـ \".board\") واستخدام خلايا Jupyter كجزء من السبورة </li> <li> <b>التكبير:</b> تغيير حجم الخط بسهولة </li> <li> <b>طي الكود:</b> رؤية هيكل المدخلات </li> <li> <b>تنسيق الكود:</b> انقر على زر التنسيق لتنسيق الكود وmarkdown تلقائيًا. </li> <li> <b>مظهر حديث:</b> أزرار، قوائم وتلميحات تنفيذ الخلايا تعكس الحالة بشكل أفضل </li> <li> <b>الإخراج الكبير:</b> التحكم في الخنق، التصفية، والتخزين المؤقت على الجانب الخادم </li> <li> <b>التقاط الخلفية للإخراج:</b> يعمل إذا لم يكن هناك مستخدم يفتح دفتر الملاحظات (<A>مناقشة</A>) </li> <li> <b>دعم الهواتف المحمولة:</b> الهواتف والأجهزة اللوحية </li> <li> <b>إنشاء الخلايا:</b> انقر على الخط الأزرق بين الخلايا لإنشاء خلايا جديدة </li> <li> <b>مشاركة:</b> عملك مرئي للجمهور عبر عارض الدفاتر الخفيف والسريع الخاص بنا </li> <li> <b>LaTeX:</b> تصدير دفتر الملاحظات إلى LaTeX، ثم تحرير LaTeX المُولد مباشرة في CoCalc. </li> <li> <b>اختصارات المفاتيح ومخططات الألوان:</b> اختصارات VIM، Emacs، وSublime، والعديد من مخططات الألوان (في إعدادات الحساب) </li> </ul>",
   "jupyter.about.message": "<Paragraph>أنت تستخدم دفتر ملاحظات CoCalc Jupyter.</Paragraph> <Paragraph> دفتر ملاحظات CoCalc Jupyter هو إعادة كتابة كاملة مفتوحة المصدر من قبل SageMath, Inc. لعميل دفتر الملاحظات الكلاسيكي من مشروع <A1>Jupyter</A1>. يحتفظ دفتر ملاحظات CoCalc Jupyter بالتوافق الكامل مع تنسيق الملفات والمظهر العام لدفتر الملاحظات الكلاسيكي. يحسن على الدفتر الكلاسيكي كما يلي: {features} بعض وظائف الإضافات الكلاسيكية لم تدعم بعد (إذا كنت بحاجة إلى شيء ما، <A2>تحقق هنا</A2> وأنشئ {support_link})، وبعض ما سبق متاح أيضًا في Jupyter الكلاسيكي عبر الإضافات. </Paragraph> <Title>أسئلة</Title> {faq} <Title>معلومات الخادم</Title> {server_info} <Title>معلومات النواة الحالية</Title> {kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "مسح المخرجات",
   "jupyter.cell-buttonbar-menu.copy-cell": "نسخ الخلية",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "نسخ إلى الحافظة",
   "jupyter.cell-buttonbar-menu.cut-cell": "قص الخلية",
