+++ src/packages/frontend/i18n/trans/de_DE.json
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "Türkisch",
   "jupyter.about.features": "<ul> <li> <b>Realtimesynchronisation:</b> Mehrere Personen können gleichzeitig Notebooks bearbeiten: mehrere Cursor, dokumentweite benutzerspezifische Rückgängig- und Wiederherstellungsfunktionen, realtime-synchronisierte ipywidgets </li> <li> <b>Dokumentaufteilung:</b> Bearbeiten und Anzeigen mehrerer Teile eines großen Notebooks gleichzeitig </li> <li> <b>TimeTravel:</b> Detaillierte Historie anzeigen, wie ein Notebook erstellt wurde </li> <li> <b>Text/Markdown:</b> Leistungsstarker grafischer Editor für Markdown-Zellen, damit Sie schön formatierten Text schreiben können, um Ihren Code zu erklären, ohne direkt Markdown schreiben zu müssen. </li> <li> <b>KI-Assistent:</b> Mehrere KI-Tools helfen Ihnen, Fehlermeldungen zu verstehen, Code zu verfeinern, neue Codezellen basierend auf dem Kontext zu generieren oder bestehenden Code zu erklären/dokumentieren. </li> <li> <b>Snippets:</b> Codebeispiele für viele Kernel </li> <li> <b>Whiteboard:</b> Erstellen Sie ein Whiteboard (Dokument endet in \".board\") und verwenden Sie Jupyter-Zellen als Teil Ihres Whiteboards </li> <li> <b>Zoom:</b> Schriftgröße einfach ändern </li> <li> <b>Code-Faltung:</b> Struktur der Eingabe anzeigen </li> <li> <b>Code-Formatierung:</b> Klicken Sie auf die Schaltfläche Formatieren, um Ihren Code und Markdown automatisch zu formatieren. </li> <li> <b>Modernes Aussehen:</b> Schaltflächen, Menüs und Zellausführungshinweise, die den Zustand besser widerspiegeln </li> <li> <b>Große Ausgabe:</b> serverseitiges Drosseln, Fenstern und Puffern </li> <li> <b>Hintergrundaufnahme von Ausgaben:</b> funktioniert, wenn kein Benutzer das Notebook geöffnet hat (<A>Diskussion</A>) </li> <li> <b>Mobile Unterstützung:</b> Telefone und Tablets </li> <li> <b>Zellen erstellen:</b> Klicken Sie auf die blaue Linie zwischen den Zellen, um neue Zellen zu erstellen </li> <li> <b>Teilen:</b> Ihre Arbeit ist öffentlich über unseren schnellen, leichten Notebook-Viewer sichtbar </li> <li> <b>LaTeX:</b> Exportieren Sie das Notebook nach LaTeX und bearbeiten Sie anschließend das generierte LaTeX direkt in CoCalc. </li> <li> <b>Tastenkombinationen und Farbschemata:</b> VIM-, Emacs- und Sublime-Tastenkombinationen sowie viele Farbschemata (in den Kontoeinstellungen) </li> </ul>",
   "jupyter.about.message": "<Paragraph>Sie verwenden das CoCalc Jupyter-Notebook.</Paragraph> <Paragraph>Das CoCalc Jupyter-Notebook ist eine vollständige Open-Source-Neuentwicklung von SageMath, Inc. des klassischen Jupyter-Notebook-Clients aus dem <A1>Jupyter-Projekt</A1>. Das CoCalc Jupyter-Notebook bleibt vollständig kompatibel mit dem Dateiformat und dem allgemeinen Look and Feel des klassischen Notebooks. Es verbessert das klassische Notebook wie folgt: {features} Einige Funktionalitäten klassischer Erweiterungen werden noch nicht unterstützt (wenn Sie etwas benötigen, <A2>sehen Sie hier nach</A2> und erstellen Sie ein {support_link}), und einige der oben genannten Funktionen sind auch im klassischen Jupyter über Erweiterungen verfügbar.</Paragraph> <Title>Fragen</Title> {faq} <Title>Serverinformationen</Title> {server_info} <Title>Aktuelle Kernel-Informationen</Title> {kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "Ausgabe löschen",
   "jupyter.cell-buttonbar-menu.copy-cell": "Zelle kopieren",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "In die Zwischenablage kopieren",
   "jupyter.cell-buttonbar-menu.cut-cell": "Zelle ausschneiden",
