+++ src/packages/frontend/i18n/trans/es_ES.json
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "Turco",
   "jupyter.about.features": "<ul> <li> <b>Sincronización en tiempo real:</b> Varias personas pueden editar cuadernos simultáneamente: múltiples cursores, deshacer y rehacer específicos de usuario en todo el documento, ipywidgets sincronizados en tiempo real </li> <li> <b>División de documento:</b> edita y ve múltiples partes de un cuaderno grande a la vez </li> <li> <b>Viaje en el tiempo:</b> ve el historial detallado de cómo se creó un cuaderno </li> <li> <b>Texto/Markdown:</b> potente editor gráfico para celdas de markdown, para que puedas escribir texto bien formateado para explicar tu código sin tener que escribir markdown directamente. </li> <li> <b>Asistente de IA:</b> Varios herramientas de IA te ayudan a entender mensajes de error, refinar código, generar nuevas celdas de código basado en el contexto, o simplemente explicar/documentar código existente. </li> <li> <b>Fragmentos:</b> ejemplos de código para muchos núcleos </li> <li> <b>Pizarra:</b> crea una pizarra (documentado con la extensión \".board\") y usa celdas de Jupyter como parte de tu pizarra </li> <li> <b>Zoom:</b> cambia fácilmente el tamaño de la fuente </li> <li> <b>Plegado de código:</b> ve la estructura de la entrada </li> <li> <b>Formateo de código:</b> haz clic en el botón Formatear para formatear automáticamente tu código y markdown. </li> <li> <b>Apariencia moderna:</b> botones, menús y pistas de ejecución de celdas que reflejan mejor el estado </li> <li> <b>Salida grande:</b> limitación, paginación y almacenamiento en el servidor </li> <li> <b>Captura de salida en segundo plano:</b> funciona si ningún usuario tiene el cuaderno abierto (<A>discusión</A>) </li> <li> <b>Soporte móvil:</b> teléfonos y tabletas </li> <li> <b>Creación de celdas:</b> haz clic en la línea azul entre celdas para crear nuevas celdas </li> <li> <b>Compartir:</b> tu trabajo es visible públicamente a través de nuestro visor de cuadernos rápido y ligero </li> <li> <b>LaTeX:</b> exporta el cuaderno a LaTeX, luego edita el LaTeX generado directamente en CoCalc. </li> <li> <b>Atajos de teclado y esquemas de color:</b> atajos de teclado VIM, Emacs y Sublime, y muchos esquemas de color (en la configuración de la cuenta) </li> </ul>",
   "jupyter.about.message": "<Paragraph>Estás usando el cuaderno Jupyter de CoCalc.</Paragraph> <Paragraph>El Cuaderno Jupyter de CoCalc es una reescritura completa de código abierto por SageMath, Inc. del cliente clásico del Cuaderno Jupyter del <A1>proyecto Jupyter</A1>. El Cuaderno Jupyter de CoCalc mantiene plena compatibilidad con el formato de archivo y la apariencia general del cuaderno clásico. Mejora el cuaderno clásico de la siguiente manera: {features} Algunas funcionalidades de las extensiones clásicas aún no son compatibles (si necesitas algo, <A2>consulta aquí</A2> y crea un {support_link}), y algunas de las anteriores también están disponibles en el Jupyter clásico mediante extensiones. </Paragraph> <Title>Preguntas</Title> {faq} <Title>Información del Servidor</Title> {server_info} <Title>Información del Kernel Actual</Title> {kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "Borrar salida",
   "jupyter.cell-buttonbar-menu.copy-cell": "Copiar Celda",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "Copiar al Portapapeles",
   "jupyter.cell-buttonbar-menu.cut-cell": "Cortar Celda",
