+++ src/packages/frontend/i18n/trans/es_PV.json
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "Turkiera",
   "jupyter.about.features": "<ul> <li> <b>Denbora errealeko sinkronizazioa:</b> Hainbat pertsonak aldi berean edita ditzakete koadernoak: hainbat kurtsore, dokumentu osoko erabiltzaileentzako desegin eta berregin, denbora errealeko sinkronizazio bidezko ipywidgetak </li> <li> <b>Dokumentua zatitzea:</b> editatu eta ikusi koaderno handi baten hainbat zati aldi berean </li> <li> <b>DenboranBidaia:</b> ikusi koaderno bat nola sortu zen zehatz-mehatz </li> <li> <b>Testua/Markdown:</b> markdown gelaxken editore grafiko indartsua, zure kodea azaltzeko testu formatu egokian idatz dezazun markdown zuzenean idatzi beharrik gabe.</li> <li> <b>AI Laguntzailea:</b> Hainbat AI tresnak errore mezuak ulertzen, kodea fintzen, testuinguruan oinarritutako kode gelaxka berriak sortzen edo dagoen kodea azaltzen/dokumentatzen laguntzen dizute. </li> <li> <b>Zatiak:</b> kode adibideak kernel askorentzat </li> <li> <b>Arbel zuria:</b> sortu arbel zuri bat (\".board\"-rekin amaitzen dena dokumentatuta) eta erabili Jupyter zelulak zure arbel zuriaren zati gisa </li> <li> <b>Zooma:</b> aldatu letra-tamaina erraz </li> <li> <b>Kodearen tolestura:</b> ikusi sarreraren egitura </li> <li> <b>Kodearen formatua:</b> egin klik Formatua botoian zure kodea eta markdown-a automatikoki formateatzeko. </li> <li> <b>Itxura modernoa:</b> egoera hobeto islatzen duten botoiak, menuak eta gelaxkak exekutatzeko aholkuak </li> <li> <b>Irteera handia:</b> zerbitzariaren aldeko mugatzea, leihoak eta bufferratzea </li> <li> <b>Irteeraren atzeko planoan ateratzea:</b> funtzionatzen du erabiltzaile batek koadernoa irekita ez badu (<A>eztabaida</A>) </li> <li> <b>Mugikorretarako laguntza:</b> telefonoak eta tabletak </li> <li> <b>Zelaulen sorrera:</b> egin klik gelaxken arteko lerro urdinean gelaxka berriak sortzeko </li> <li> <b>Partekatu:</b> zure lana publikoki ikusgai dago gure koaderno ikustaile arin eta azkarraren bidez </li> <li> <b>LaTeX:</b> esportatu koadernoa LaTeX-era, eta editatu sortutako LaTeX zuzenean CoCalc-en. </li> <li> <b>Tekla-konbinazioak eta kolore-eskemak:</b> VIM, Emacs eta Sublime tekla-konbinazioak, eta kolore-eskema asko (kontuaren ezarpenetan) </li> </ul>",
   "jupyter.about.message": "<Paragraph>CoCalc Jupyter koadernoa erabiltzen ari zara.</Paragraph> <Paragraph> CoCalc Jupyter Notebook SageMath, Inc.-ek Jupyter proiektuaren <A1>Jupyter proiektutik</A1> klasikoaren Jupyter Notebook bezeroaren berridazketa osoa da. CoCalc Jupyter Notebook-ek fitxategi formatua eta koaderno klasikoaren itxura orokorra guztiz bateragarriak mantentzen ditu. Koaderno klasikoa hobetzen du honako hauetan: {features} Klasikoen luzapenetako funtzionalitate batzuk oraindik ez dira onartzen (zerbait behar baduzu, <A2>hemen begiratu</A2> eta sortu {support_link}), eta goiko batzuk ere Jupyter klasikoan eskuragarri daude luzapenen bidez. </Paragraph> <Title>Galderak</Title> {faq} <Title>Zerbitzariaren informazioa</Title> {server_info} <Title>Uneko Kernelaren informazioa</Title> {kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "Emaitza Garbitu",
   "jupyter.cell-buttonbar-menu.copy-cell": "Kopiatu Zelula",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "Arbelera kopiatu",
   "jupyter.cell-buttonbar-menu.cut-cell": "Moztu Zela",
