+++ src/packages/frontend/i18n/trans/fr_FR.json
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "Turc",
   "jupyter.about.features": "<ul> <li> <b>Sync en temps réel :</b> Plusieurs personnes peuvent éditer des notebooks simultanément : multiples curseurs, annulation et rétablissement spécifiques à l'utilisateur sur l'ensemble du document, ipywidgets synchronisés en temps réel </li> <li> <b>Division du document :</b> éditer et voir plusieurs parties d'un grand notebook en même temps </li> <li> <b>TimeTravel :</b> voir l'historique détaillé de la création d'un notebook </li> <li> <b>Texte/Markdown :</b> éditeur graphique puissant pour les cellules markdown, afin que vous puissiez écrire du texte bien formaté pour expliquer votre code sans avoir à écrire directement en markdown. </li> <li> <b>Assistant AI :</b> Plusieurs outils AI vous aident à comprendre les messages d'erreur, affiner le code, générer de nouvelles cellules de code en fonction du contexte, ou simplement expliquer/documenter le code existant. </li> <li> <b>Extraits :</b> exemples de code pour de nombreux noyaux </li> <li> <b>Tableau blanc :</b> créer un tableau blanc (document se terminant par \".board\") et utiliser les cellules Jupyter comme partie de votre tableau blanc </li> <li> <b>Zoom :</b> changer facilement la taille de la police </li> <li> <b>Repliement de code :</b> voir la structure de l'entrée </li> <li> <b>Formatage de code :</b> cliquer sur le bouton Format pour formater automatiquement votre code et markdown. </li> <li> <b>Apparence moderne :</b> boutons, menus et indices d'exécution des cellules qui reflètent mieux l'état </li> <li> <b>Grande sortie :</b> limitation, fenêtrage et mise en mémoire tampon côté serveur </li> <li> <b>Capture de sortie en arrière-plan :</b> fonctionne si aucun utilisateur n'a ouvert le notebook (<A>discussion</A>) </li> <li> <b>Support mobile :</b> téléphones et tablettes </li> <li> <b>Création de cellules :</b> cliquer sur la ligne bleue entre les cellules pour créer de nouvelles cellules </li> <li> <b>Partager :</b> votre travail est visible publiquement via notre visualiseur de notebook rapide et léger </li> <li> <b>LaTeX :</b> exporter le notebook en LaTeX, puis éditer le LaTeX généré directement dans CoCalc. </li> <li> <b>Raccourcis clavier et schémas de couleurs :</b> raccourcis clavier VIM, Emacs et Sublime, et de nombreux schémas de couleurs (dans les paramètres du compte) </li> </ul>",
   "jupyter.about.message": "<Paragraph>Vous utilisez le notebook Jupyter de CoCalc.</Paragraph> <Paragraph>Le notebook Jupyter de CoCalc est une réécriture complète en open source par SageMath, Inc. du client classique du notebook Jupyter du <A1>projet Jupyter</A1>. Le notebook Jupyter de CoCalc maintient une compatibilité totale avec le format de fichier et l'apparence générale du notebook classique. Il améliore le notebook classique de la manière suivante : {features} Certaines fonctionnalités des extensions classiques ne sont pas encore supportées (si vous avez besoin de quelque chose, <A2>vérifiez ici</A2> et créez un {support_link}), et certaines des améliorations ci-dessus sont également disponibles dans le Jupyter classique via des extensions. </Paragraph> <Title>Questions</Title> {faq} <Title>Informations sur le serveur</Title> {server_info} <Title>Informations sur le noyau actuel</Title> {kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "Effacer la sortie",
   "jupyter.cell-buttonbar-menu.copy-cell": "Copier la cellule",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "Copier dans le Presse-papiers",
   "jupyter.cell-buttonbar-menu.cut-cell": "Couper la cellule",
