+++ src/packages/frontend/i18n/trans/he_IL.json
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "טורקית",
   "jupyter.about.features": "<ul> <li> <b>סנכרון בזמן אמת:</b> מספר אנשים יכולים לערוך מחברות בו זמנית: מספר סמנים, ביטול ושחזור ספציפיים למשתמש על גבי כל המסמך, סנכרון בזמן אמת של ipywidgets </li> <li> <b>פיצול מסמך:</b> לערוך ולראות חלקים מרובים של מחברת גדולה בבת אחת </li> <li> <b>TimeTravel:</b> לראות היסטוריה מפורטת של איך בדיוק נוצרה המחברת </li> <li> <b>טקסט/Markdown:</b> עורך גרפי חזק לתאי markdown, כך שתוכל לכתוב טקסט מעוצב יפה להסביר את הקוד שלך בלי לכתוב markdown ישירות. </li> <li> <b>עוזר AI:</b> מספר כלים של AI עוזרים לך להבין הודעות שגיאה, לשפר קוד, לייצר תאי קוד חדשים בהתבסס על הקשר, או פשוט להסביר/לתעד קוד קיים. </li> <li> <b>Snippets:</b> דוגמאות קוד להרבה ליבות </li> <li> <b>לוח מחיק:</b> צור לוח מחיק (מסמך עם סיומת \".board\") והשתמש בתאי Jupyter כחלק מהלוח המחיק שלך </li> <li> <b>Zoom:</b> לשנות בקלות את גודל הגופן </li> <li> <b>קיפול קוד:</b> לראות את מבנה הקלט </li> <li> <b>עיצוב קוד:</b> לחץ על כפתור העיצוב כדי לעצב אוטומטית את הקוד וה-markdown שלך. </li> <li> <b>מראה מודרני:</b> כפתורים, תפריטים ורמזים להרצת תאים שמשקפים את המצב בצורה טובה יותר </li> <li> <b>פלט גדול:</b> הגבלת תעבורה, חלונות וזיכרון בצד השרת </li> <li> <b>תפיסת פלט ברקע:</b> עובד אם אף משתמש לא פתח את המחברת (<A>דיון</A>) </li> <li> <b>תמיכה במובייל:</b> טלפונים וטאבלטים </li> <li> <b>יצירת תאים:</b> לחץ על הקו הכחול בין התאים כדי ליצור תאים חדשים </li> <li> <b>שיתוף:</b> העבודה שלך גלויה לציבור דרך צופה מחברות מהיר וקל </li> <li> <b>LaTeX:</b> יצא מחברת ל-LaTeX, ואז ערוך את ה-LaTeX שנוצר ישירות ב-CoCalc. </li> <li> <b>קיצורי מקלדת וסכמות צבע:</b> קיצורי מקלדת של VIM, Emacs ו-Sublime, והרבה סכמות צבע (בהגדרות החשבון) </li> </ul>",
   "jupyter.about.message": "<Paragraph>אתה משתמש בפנקס הרשימות של CoCalc Jupyter.</Paragraph> <Paragraph> פנקס הרשימות של CoCalc Jupyter הוא שכתוב מלא בקוד פתוח על ידי SageMath, Inc. של לקוח פנקס הרשימות הקלאסי מה<A1>פרויקט Jupyter</A1>. פנקס הרשימות של CoCalc Jupyter שומר על תאימות מלאה עם פורמט הקבצים והמראה הכללי של הפנקס הקלאסי. הוא משפר את הפנקס הקלאסי באופן הבא: {features} חלק מהפונקציות של הרחבות קלאסיות עדיין לא נתמכות (אם אתה צריך משהו, <A2>בדוק כאן</A2> וצור {support_link}), וחלק מהנ\"ל זמין גם ב-Jupyter הקלאסי דרך הרחבות. </Paragraph> <Title>שאלות</Title> {faq} <Title>מידע על השרת</Title> {server_info} <Title>מידע על הליבה הנוכחית</Title> {kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "נקה פלט",
   "jupyter.cell-buttonbar-menu.copy-cell": "העתק תא",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "העתק ללוח",
   "jupyter.cell-buttonbar-menu.cut-cell": "גזור תא",
