+++ src/packages/frontend/i18n/trans/hi_IN.json
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "तुर्की",
   "jupyter.about.features": "<ul> <li> <b>रीयलटाइम सिंक:</b> कई लोग एक साथ नोटबुक संपादित कर सकते हैं: कई कर्सर, दस्तावेज़-व्यापी उपयोगकर्ता-विशिष्ट पूर्ववत और पुनः करें, रीयलटाइम सिंक्रनाइज़्ड ipywidgets </li> <li> <b>दस्तावेज़ विभाजन:</b> एक बड़े नोटबुक के कई हिस्सों को एक साथ संपादित और देख सकते हैं </li> <li> <b>टाइमट्रैवल:</b> यह देखें कि एक नोटबुक कैसे बनाई गई थी इसका विस्तृत इतिहास </li> <li> <b>पाठ/मार्कडाउन:</b> मार्कडाउन कोशिकाओं के लिए शक्तिशाली ग्राफिकल संपादक, ताकि आप सीधे मार्कडाउन लिखे बिना अपने कोड को समझाने के लिए अच्छे से स्वरूपित पाठ लिख सकें </li> <li> <b>एआई सहायक:</b> कई एआई उपकरण आपको त्रुटि संदेशों को समझने, कोड को परिष्कृत करने, संदर्भ के आधार पर नई कोड कोशिकाएं उत्पन्न करने, या बस मौजूदा कोड को समझाने/दस्तावेज़ीकरण करने में मदद करते हैं </li> <li> <b>स्निपेट्स:</b> कई कर्नल्स के लिए कोड नमूने </li> <li> <b>व्हाइटबोर्ड:</b> एक व्हाइटबोर्ड (\".board\" में समाप्त होने वाला दस्तावेज़) बनाएं और अपने व्हाइटबोर्ड का हिस्सा बनने के लिए जुपिटर कोशिकाओं का उपयोग करें </li> <li> <b>जूम:</b> आसानी से फ़ॉन्ट आकार बदलें </li> <li> <b>कोड फोल्डिंग:</b> इनपुट की संरचना देखें </li> <li> <b>कोड स्वरूपण:</b> अपने कोड और मार्कडाउन को स्वचालित रूप से स्वरूपित करने के लिए स्वरूप बटन पर क्लिक करें </li> <li> <b>आधुनिक रूप:</b> बटन, मेनू और सेल निष्पादन संकेत जो राज्य को बेहतर तरीके से दर्शाते हैं </li> <li> <b>बड़ा आउटपुट:</b> सर्वर-साइड थ्रॉटलिंग, विंडोइंग, और बफरिंग </li> <li> <b>पृष्ठभूमि में आउटपुट कैप्चर:</b> काम करता है यदि कोई उपयोगकर्ता नोटबुक खोलकर नहीं रखता है (<A>चर्चा</A>) </li> <li> <b>मोबाइल समर्थन:</b> फोन और टैबलेट </li> <li> <b>कोशिका निर्माण:</b> नई कोशिकाएं बनाने के लिए कोशिकाओं के बीच नीली रेखा पर क्लिक करें </li> <li> <b>साझा करें:</b> आपका कार्य हमारे तेज़ हल्के नोटबुक व्यूअर के माध्यम से सार्वजनिक रूप से दिखाई देता है </li> <li> <b>LaTeX:</b> नोटबुक को LaTeX में निर्यात करें, फिर उत्पन्न LaTeX को सीधे CoCalc में संपादित करें </li> <li> <b>कीबाइंडिंग्स और रंग योजनाएं:</b> VIM, Emacs, और Sublime कीबाइंडिंग्स, और कई रंग योजनाएं (खाता सेटिंग्स में) </li> </ul>",
   "jupyter.about.message": "आप CoCalc Jupyter नोटबुक का उपयोग कर रहे हैं CoCalc Jupyter Notebook एक पूर्ण ओपन सोर्स पुनर्लेखन है SageMath, Inc. द्वारा क्लासिकल Jupyter Notebook क्लाइंट का, जो <A1>Jupyter प्रोजेक्ट</A1> का हिस्सा है। CoCalc Jupyter Notebook क्लासिकल नोटबुक के फाइल फॉर्मेट और सामान्य लुक और फील के साथ पूरी तरह संगत है। यह क्लासिकल नोटबुक में निम्नलिखित सुधार करता है: {features} क्लासिकल एक्सटेंशन्स की कुछ कार्यक्षमता अभी तक समर्थित नहीं है (यदि आपको कुछ चाहिए, तो <A2>यहाँ देखें</A2> और एक {support_link} बनाएं), और उपरोक्त में से कुछ क्लासिकल Jupyter में एक्सटेंशन्स के माध्यम से भी उपलब्ध है। प्रश्न {faq} सर्वर जानकारी {server_info} वर्तमान कर्नेल जानकारी {kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "आउटपुट साफ़ करें",
   "jupyter.cell-buttonbar-menu.copy-cell": "सेल कॉपी करें",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "क्लिपबोर्ड में कॉपी करें",
   "jupyter.cell-buttonbar-menu.cut-cell": "कोशिका काटें",
