+++ src/packages/frontend/i18n/trans/hu_HU.json
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "Török",
   "jupyter.about.features": "<ul> <li> <b>Valós idejű szinkronizálás:</b> Több ember egyszerre szerkesztheti a jegyzetfüzeteket: több kurzor, dokumentum-szintű felhasználóspecifikus visszavonás és újra, valós idejű szinkronizált ipywidgets </li> <li> <b>Dokumentum felosztása:</b> egyszerre szerkeszthet és láthat több részt egy nagy jegyzetfüzetből </li> <li> <b>Időutazás:</b> részletes történelmet láthat arról, hogyan készült a jegyzetfüzet </li> <li> <b>Szöveg/Markdown:</b> erőteljes grafikus szerkesztő markdown cellákhoz, így szépen formázott szöveget írhat a kód magyarázatához anélkül, hogy közvetlenül markdown-t kellene írnia </li> <li> <b>AI Asszisztens:</b> Több AI eszköz segít megérteni a hibaüzeneteket, finomítani a kódot, új kódcellákat generálni a kontextus alapján, vagy egyszerűen csak magyarázni/dokumentálni a meglévő kódot. </li> <li> <b>Kódrészletek:</b> kódminták sok kernelhez </li> <li> <b>Tábla:</b> hozzon létre egy táblát (dokumentum \".board\" végződéssel) és használja a Jupyter cellákat a tábla részeként </li> <li> <b>Nagyítás:</b> könnyen változtatható betűméret </li> <li> <b>Kód összehajtás:</b> látható a bevitel szerkezete </li> <li> <b>Kód formázás:</b> kattintson a Formázás gombra a kód és markdown automatikus formázásához </li> <li> <b>Modern kinézet:</b> gombok, menük és cellavégrehajtási tippek, amelyek jobban tükrözik az állapotot </li> <li> <b>Nagy kimenet:</b> szerver-oldali korlátozás, ablakolás és pufferelés </li> <li> <b>Háttérkimenet rögzítése:</b> működik, ha senki nem nyitotta meg a jegyzetfüzetet (<A>beszélgetés</A>) </li> <li> <b>Mobil támogatás:</b> telefonok és tabletek </li> <li> <b>Cella létrehozása:</b> kattintson a kék vonalra a cellák között új cellák létrehozásához </li> <li> <b>Megosztás:</b> munkája nyilvánosan látható a gyors és könnyű jegyzetfüzet megtekintőnkön keresztül </li> <li> <b>LaTeX:</b> exportálja a jegyzetfüzetet LaTeX-be, majd szerkessze a generált LaTeX-et közvetlenül a CoCalc-ban </li> <li> <b>Billentyűparancsok és színsémák:</b> VIM, Emacs és Sublime billentyűparancsok, és sok színséma (a fiók beállításainál) </li> </ul>",
   "jupyter.about.message": "<Paragraph>Ön a CoCalc Jupyter notebookot használja.</Paragraph> <Paragraph>A CoCalc Jupyter Notebook a SageMath, Inc. teljes nyílt forráskódú újraírása a klasszikus Jupyter Notebook kliensnek a <A1>Jupyter projektből</A1>. A CoCalc Jupyter Notebook teljes mértékben kompatibilis a fájlformátummal és a klasszikus notebook általános kinézetével és érzésével. A következő módokon javítja a klasszikus notebookot: {features} Néhány klasszikus bővítmény funkciója még nem támogatott (ha szüksége van valamire, <A2>ellenőrizze itt</A2> és hozzon létre egy {support_link}), és néhány fentiek közül a klasszikus Jupyterben is elérhető kiterjesztéseken keresztül.</Paragraph> <Title>Kérdések</Title> {faq} <Title>Szerverinformációk</Title> {server_info} <Title>Jelenlegi Kernel Információk</Title> {kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "Kimenet törlése",
   "jupyter.cell-buttonbar-menu.copy-cell": "Másolja a cellát",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "Másolás a vágólapra",
   "jupyter.cell-buttonbar-menu.cut-cell": "Kivágás cella",
