+++ src/packages/frontend/i18n/trans/it_IT.json
@@ -597,7 +597,7 @@
   "editor.snippets.label": "Snippet",
   "editor.snippets.tooltip": "Apri un pannello contenente frammenti di codice",
   "editor.table_of_contents.name": "Indice",
-  "editor.table_of_contents.short": "Contenuto",
+  "editor.table_of_contents.short": "Contenuti",
   "editor.terminal.cmd.help.title": "Mostra la documentazione per l'utilizzo del Terminale Linux in CoCalc.",
   "file_actions.compress.name": "Comprimi",
   "file_actions.copy.name": "Copia",
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "Turco",
   "jupyter.about.features": "<ul> <li> <b>Sincronizzazione in tempo reale:</b> Più persone possono modificare simultaneamente i notebook: più cursori, annullamento e ripetizione specifici per utente su tutto il documento, ipywidgets sincronizzati in tempo reale </li> <li> <b>Divisione del documento:</b> modifica e visualizza più parti di un grande notebook contemporaneamente </li> <li> <b>TimeTravel:</b> visualizza la cronologia dettagliata di come è stato creato un notebook </li> <li> <b>Testo/Markdown:</b> potente editor grafico per le celle markdown, così puoi scrivere testo ben formattato per spiegare il tuo codice senza dover scrivere direttamente in markdown. </li> <li> <b>Assistente AI:</b> Diversi strumenti AI ti aiutano a comprendere i messaggi di errore, perfezionare il codice, generare nuove celle di codice in base al contesto o semplicemente spiegare/documentare il codice esistente. </li> <li> <b>Snippet:</b> esempi di codice per molti kernel </li> <li> <b>Lavagna:</b> crea una lavagna (documento che termina con \".board\") e usa le celle Jupyter come parte della tua lavagna </li> <li> <b>Zoom:</b> cambia facilmente la dimensione del carattere </li> <li> <b>Racchiudere codice:</b> visualizza la struttura dell'input </li> <li> <b>Formattazione del codice:</b> clicca sul pulsante Format per formattare automaticamente il tuo codice e markdown. </li> <li> <b>Aspetto moderno:</b> pulsanti, menu e suggerimenti di esecuzione delle celle che riflettono meglio lo stato </li> <li> <b>Uscita ampia:</b> limitazione, finestratura e buffering lato server </li> <li> <b>Cattura del output in background:</b> funziona se nessun utente ha il notebook aperto (<A>discussione</A>) </li> <li> <b>Supporto mobile:</b> telefoni e tablet </li> <li> <b>Creazione di celle:</b> clicca sulla linea blu tra le celle per creare nuove celle </li> <li> <b>Condividi:</b> il tuo lavoro è visibile pubblicamente tramite il nostro visualizzatore di notebook leggero e veloce </li> <li> <b>LaTeX:</b> esporta il notebook in LaTeX, poi modifica il LaTeX generato direttamente in CoCalc. </li> <li> <b>Combinazioni di tasti e schemi di colore:</b> Combinazioni di tasti VIM, Emacs e Sublime, e molti schemi di colore (nelle impostazioni dell'account) </li> </ul>",
   "jupyter.about.message": "<Paragraph>Stai usando il CoCalc Jupyter notebook.</Paragraph> <Paragraph>CoCalc Jupyter Notebook è una riscrittura completa open source da parte di SageMath, Inc. del client classico Jupyter Notebook del <A1>progetto Jupyter</A1>. CoCalc Jupyter Notebook mantiene la piena compatibilità con il formato dei file e l'aspetto generale del notebook classico. Migliora il notebook classico come segue: {features} Alcune funzionalità delle estensioni classiche non sono ancora supportate (se hai bisogno di qualcosa, <A2>controlla qui</A2> e crea un {support_link}), e alcune delle suddette funzionalità sono anche disponibili nel Jupyter classico tramite estensioni.</Paragraph> <Title>Domande</Title> {faq} <Title>Informazioni sul Server</Title> {server_info} <Title>Informazioni sul Kernel Attuale</Title> {kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "Cancella Output",
   "jupyter.cell-buttonbar-menu.copy-cell": "Copia cella",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "Copia negli appunti",
   "jupyter.cell-buttonbar-menu.cut-cell": "Taglia Cellula",
