+++ src/packages/frontend/i18n/trans/ja_JP.json
@@ -597,7 +597,7 @@
   "editor.snippets.label": "スニペット",
   "editor.snippets.tooltip": "コードスニペットを含むパネルを開く",
   "editor.table_of_contents.name": "目次",
-  "editor.table_of_contents.short": "コンテンツ",
+  "editor.table_of_contents.short": "目次",
   "editor.terminal.cmd.help.title": "CoCalcでLinuxターミナルを使用するためのドキュメントを表示。",
   "file_actions.compress.name": "圧縮",
   "file_actions.copy.name": "コピー",
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "トルコ語",
   "jupyter.about.features": "<ul> <li> <b>リアルタイム同期:</b> 複数の人が同時にノートブックを編集できる: 複数のカーソル、ドキュメント全体のユーザー特有の元に戻すとやり直し、リアルタイム同期されたipywidgets </li> <li> <b>ドキュメント分割:</b> 大きなノートブックの複数の部分を同時に編集および表示 </li> <li> <b>タイムトラベル:</b> ノートブックがどのように作成されたかの詳細な履歴を確認 </li> <li> <b>テキスト/Markdown:</b> マークダウンセルの強力なグラフィカルエディタ、コードを直接マークダウンを書くことなく、きれいにフォーマットされたテキストで説明できます </li> <li> <b>AIアシスタント:</b> 複数のAIツールがエラーメッセージの理解、コードの改善、コンテキストに基づく新しいコードセルの生成、または既存のコードの説明/文書化を支援します </li> <li> <b>スニペット:</b> 多くのカーネルのコードサンプル </li> <li> <b>ホワイトボード:</b> ホワイトボード（\".board\"で終わるドキュメント）を作成し、Jupyterセルをホワイトボードの一部として使用 </li> <li> <b>ズーム:</b> フォントサイズを簡単に変更 </li> <li> <b>コード折りたたみ:</b> 入力の構造を確認 </li> <li> <b>コードフォーマット:</b> フォーマットボタンをクリックして、コードとマークダウンを自動的にフォーマット </li> <li> <b>モダンな外観:</b> 状態をよりよく反映するボタン、メニュー、およびセル実行のヒント </li> <li> <b>大きな出力:</b> サーバー側のスロットリング、ウィンドウ化、およびバッファリング </li> <li> <b>出力のバックグラウンドキャプチャ:</b> ユーザーがノートブックを開いていない場合に機能します (<A>ディスカッション</A>) </li> <li> <b>モバイル対応:</b> 携帯電話とタブレット </li> <li> <b>セル作成:</b> セル間の青い線をクリックして新しいセルを作成 </li> <li> <b>共有:</b> あなたの作業は、私たちの高速で軽量なノートブックビューアを通じて公開されます </li> <li> <b>LaTeX:</b> ノートブックをLaTeXにエクスポートし、CoCalcで生成されたLaTeXを直接編集 </li> <li> <b>キーバインドとカラースキーム:</b> VIM、Emacs、およびSublimeのキーバインド、アカウント設定で多くのカラースキーム </li> </ul>",
   "jupyter.about.message": "<Paragraph>CoCalc Jupyterノートブックを使用しています。</Paragraph> <Paragraph>CoCalc Jupyterノートブックは、<A1>Jupyterプロジェクト</A1>の古典的なJupyterノートブッククライアントをSageMath, Inc.が完全にオープンソースで書き直したものです。CoCalc Jupyterノートブックは、ファイル形式と古典的なノートブックの一般的な外観と操作感との完全な互換性を保っています。次のように古典的なノートブックを改良しています：{features} 古典的な拡張機能の一部の機能はまだサポートされていません（必要なものがある場合は、<A2>こちらを確認</A2>して{support_link}を作成してください）。また、上記の一部は古典的なJupyterでも拡張機能を通じて利用可能です。</Paragraph> <Title>質問</Title> {faq} <Title>サーバー情報</Title> {server_info} <Title>現在のカーネル情報</Title> {kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "出力をクリア",
   "jupyter.cell-buttonbar-menu.copy-cell": "セルをコピー",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "クリップボードにコピー",
   "jupyter.cell-buttonbar-menu.cut-cell": "セルを切り取る",
