+++ src/packages/frontend/i18n/trans/ko_KR.json
@@ -597,7 +597,7 @@
   "editor.snippets.label": "스니펫",
   "editor.snippets.tooltip": "코드 스니펫을 포함하는 패널 열기",
   "editor.table_of_contents.name": "목차",
-  "editor.table_of_contents.short": "내용",
+  "editor.table_of_contents.short": "목차",
   "editor.terminal.cmd.help.title": "CoCalc에서 Linux 터미널 사용 설명서 보기",
   "file_actions.compress.name": "압축",
   "file_actions.copy.name": "복사",
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "터키어",
   "jupyter.about.features": "<ul> <li> <b>실시간 동기화:</b> 여러 사람이 동시에 노트북을 편집할 수 있습니다: 여러 커서, 문서 전체 사용자별 실행 취소 및 다시 실행, 실시간 동기화된 ipywidgets </li> <li> <b>문서 분할:</b> 큰 노트북의 여러 부분을 한 번에 편집하고 볼 수 있습니다 </li> <li> <b>TimeTravel:</b> 노트북이 어떻게 생성되었는지 상세한 기록을 볼 수 있습니다 </li> <li> <b>텍스트/마크다운:</b> 마크다운 셀을 위한 강력한 그래픽 편집기로, 마크다운을 직접 작성하지 않고도 코드를 설명하는 멋지게 서식화된 텍스트를 작성할 수 있습니다 </li> <li> <b>AI 어시스턴트:</b> 여러 AI 도구가 오류 메시지를 이해하고, 코드를 정제하며, 상황에 맞는 새로운 코드 셀을 생성하거나 기존 코드를 설명/문서화하는 데 도움을 줍니다 </li> <li> <b>스니펫:</b> 많은 커널에 대한 코드 샘플 </li> <li> <b>화이트보드:</b> 화이트보드(“.board”로 끝나는 문서)를 생성하고 Jupyter 셀을 화이트보드의 일부로 사용할 수 있습니다 </li> <li> <b>줌:</b> 글꼴 크기를 쉽게 변경할 수 있습니다 </li> <li> <b>코드 접기:</b> 입력 구조를 볼 수 있습니다 </li> <li> <b>코드 서식:</b> 포맷 버튼을 클릭하여 코드와 마크다운을 자동으로 서식화합니다 </li> <li> <b>현대적 외관:</b> 상태를 더 잘 반영하는 버튼, 메뉴 및 셀 실행 힌트 </li> <li> <b>대용량 출력:</b> 서버 측 스로틀링, 창 설정 및 버퍼링 </li> <li> <b>출력의 백그라운드 캡처:</b> 사용자가 노트북을 열지 않아도 작동합니다 (<A>토론</A>) </li> <li> <b>모바일 지원:</b> 전화기 및 태블릿 </li> <li> <b>셀 생성:</b> 셀 사이의 파란 선을 클릭하여 새 셀을 생성합니다 </li> <li> <b>공유:</b> 우리의 빠르고 가벼운 노트북 뷰어를 통해 작업이 공개적으로 보입니다 </li> <li> <b>LaTeX:</b> 노트북을 LaTeX로 내보낸 후 생성된 LaTeX를 CoCalc에서 직접 편집할 수 있습니다 </li> <li> <b>키바인딩 및 색 구성표:</b> VIM, Emacs, Sublime 키바인딩 및 다양한 색 구성표 (계정 설정에서) </li> </ul>",
   "jupyter.about.message": "<Paragraph>당신은 CoCalc Jupyter 노트북을 사용하고 있습니다.</Paragraph> <Paragraph> CoCalc Jupyter Notebook은 SageMath, Inc.가 <A1>Jupyter 프로젝트</A1>의 클래식 Jupyter Notebook 클라이언트를 완전히 오픈 소스로 다시 작성한 것입니다. CoCalc Jupyter Notebook은 파일 형식과 일반적인 모양 및 느낌에서 클래식 노트북과 완벽한 호환성을 유지합니다. 다음과 같이 클래식 노트북을 개선했습니다: {features} 클래식 확장의 일부 기능은 아직 지원되지 않습니다 (필요한 경우, <A2>여기를 확인하세요</A2> 그리고 {support_link}를 생성하세요), 위의 일부는 확장을 통해 클래식 Jupyter에서도 사용할 수 있습니다. </Paragraph> <Title>질문</Title> {faq} <Title>서버 정보</Title> {server_info} <Title>현재 커널 정보</Title> {kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "출력 지우기",
   "jupyter.cell-buttonbar-menu.copy-cell": "셀 복사",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "클립보드에 복사",
   "jupyter.cell-buttonbar-menu.cut-cell": "셀 잘라내기",
