+++ src/packages/frontend/i18n/trans/nl_NL.json
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "Turks",
   "jupyter.about.features": "<ul> <li> <b>Realtime sync:</b> Meerdere mensen kunnen tegelijkertijd notities bewerken: meerdere cursors, documentbrede gebruiker-specifieke ongedaan maken en opnieuw doen, realtime gesynchroniseerde ipywidgets </li> <li> <b>Document splitsen:</b> bewerk en bekijk meerdere delen van een groot notitieboek tegelijk </li> <li> <b>TimeTravel:</b> zie gedetailleerde geschiedenis van precies hoe een notitieboek is gemaakt </li> <li> <b>Tekst/Markdown:</b> krachtige grafische editor voor markdown-cellen, zodat je netjes opgemaakte tekst kunt schrijven om je code uit te leggen zonder direct markdown te hoeven schrijven. </li> <li> <b>AI-assistent:</b> Verschillende AI-tools helpen je om foutmeldingen te begrijpen, code te verfijnen, nieuwe codecellen te genereren op basis van context, of gewoon bestaande code uit te leggen/documenteren. </li> <li> <b>Snippets:</b> codevoorbeelden voor veel kernels </li> <li> <b>Whiteboard:</b> maak een whiteboard (document eindigend op \".board\") en gebruik Jupyter-cellen als onderdeel van je whiteboard </li> <li> <b>Zoom:</b> gemakkelijk lettergrootte wijzigen </li> <li> <b>Code folding:</b> zie structuur van invoer </li> <li> <b>Code formatteren:</b> klik op de Format-knop om je code en markdown automatisch te formatteren. </li> <li> <b>Moderne uitstraling:</b> knoppen, menu's en celuitvoeringshints die de status beter weerspiegelen </li> <li> <b>Grote output:</b> server-side throttling, venstering en buffering </li> <li> <b>Achtergrond vastleggen van output:</b> werkt als geen gebruiker het notitieboek geopend heeft (<A>discussie</A>) </li> <li> <b>Mobiele ondersteuning:</b> telefoons en tablets </li> <li> <b>Celcreatie:</b> klik op de blauwe lijn tussen cellen om nieuwe cellen te maken </li> <li> <b>Delen:</b> je werk is openbaar zichtbaar via onze snelle, lichte notitieboekviewer </li> <li> <b>LaTeX:</b> exporteer notitieboek naar LaTeX, bewerk vervolgens de gegenereerde LaTeX direct in CoCalc. </li> <li> <b>Toetscombinaties en kleurenschema's:</b> VIM, Emacs, en Sublime toetscombinaties, en veel kleurenschema's (in accountinstellingen) </li> </ul>",
   "jupyter.about.message": "<Paragraph>U gebruikt de CoCalc Jupyter-notebook.</Paragraph> <Paragraph>CoCalc Jupyter Notebook is een complete open source herschrijving door SageMath, Inc. van de klassieke Jupyter Notebook-client van het <A1>Jupyter-project</A1>. CoCalc Jupyter Notebook behoudt volledige compatibiliteit met het bestandsformaat en de algemene look en feel van de klassieke notebook. Het verbetert de klassieke notebook als volgt: {features} Sommige functionaliteit van klassieke extensies wordt nog niet ondersteund (als u iets nodig heeft, <A2>check hier</A2> en maak een {support_link}), en sommige van de bovenstaande functies zijn ook beschikbaar in klassieke Jupyter via extensies.</Paragraph> <Title>Vragen</Title> {faq} <Title>Serverinformatie</Title> {server_info} <Title>Huidige Kernelinformatie</Title> {kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "Wis Output",
   "jupyter.cell-buttonbar-menu.copy-cell": "Kopieer cel",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "Kopiëren naar Klembord",
   "jupyter.cell-buttonbar-menu.cut-cell": "Cell Knippen",
