+++ src/packages/frontend/i18n/trans/pl_PL.json
@@ -597,7 +597,7 @@
   "editor.snippets.label": "Fragmenty",
   "editor.snippets.tooltip": "Otwórz panel zawierający fragmenty kodu",
   "editor.table_of_contents.name": "Spis treści",
-  "editor.table_of_contents.short": "Zawartość",
+  "editor.table_of_contents.short": "Spis treści",
   "editor.terminal.cmd.help.title": "Pokaż dokumentację dotyczącą używania terminala Linux w CoCalc",
   "file_actions.compress.name": "Kompresuj",
   "file_actions.copy.name": "Kopiuj",
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "Turecki",
   "jupyter.about.features": "<ul> <li> <b>Synchronizacja w czasie rzeczywistym:</b> Kilka osób może jednocześnie edytować notatniki: wiele kursorów, cofanie i ponawianie specyficzne dla użytkownika na całym dokumencie, synchronizowane w czasie rzeczywistym ipywidgets </li> <li> <b>Podział dokumentu:</b> edytuj i oglądaj kilka części dużego notatnika jednocześnie </li> <li> <b>TimeTravel:</b> zobacz szczegółową historię, jak dokładnie notatnik został utworzony </li> <li> <b>Tekst/Markdown:</b> potężny graficzny edytor dla komórek markdown, dzięki czemu możesz pisać dobrze sformatowany tekst, aby wyjaśnić swój kod bez konieczności pisania bezpośrednio w markdown. </li> <li> <b>Asystent AI:</b> Kilka narzędzi AI pomaga zrozumieć komunikaty o błędach, udoskonalić kod, generować nowe komórki kodu na podstawie kontekstu lub po prostu wyjaśniać/dokumentować istniejący kod. </li> <li> <b>Fragmenty:</b> przykłady kodu dla wielu jąder </li> <li> <b>Tablica:</b> utwórz tablicę (dokument kończący się na \".board\") i używaj komórek Jupyter jako części swojej tablicy </li> <li> <b>Powiększenie:</b> łatwo zmień rozmiar czcionki </li> <li> <b>Zwijanie kodu:</b> zobacz strukturę wejścia </li> <li> <b>Formatowanie kodu:</b> kliknij przycisk Format, aby automatycznie sformatować swój kod i markdown. </li> <li> <b>Nowoczesny wygląd:</b> przyciski, menu i wskazówki dotyczące wykonania komórek, które lepiej odzwierciedlają stan </li> <li> <b>Duży wynik:</b> ograniczanie po stronie serwera, okienkowanie i buforowanie </li> <li> <b>Tło przechwytywania wyników:</b> działa, jeśli żaden użytkownik nie ma otwartego notatnika (<A>dyskusja</A>) </li> <li> <b>Wsparcie mobilne:</b> telefony i tablety </li> <li> <b>Tworzenie komórek:</b> kliknij niebieską linię między komórkami, aby utworzyć nowe komórki </li> <li> <b>Udostępnij:</b> Twoja praca jest widoczna publicznie za pomocą naszego szybkiego i lekkiego przeglądarki notatników </li> <li> <b>LaTeX:</b> eksportuj notatnik do LaTeX, a następnie edytuj wygenerowany LaTeX bezpośrednio w CoCalc. </li> <li> <b>Skróty klawiszowe i schematy kolorów:</b> Skróty klawiszowe VIM, Emacs i Sublime oraz wiele schematów kolorów (w ustawieniach konta) </li> </ul>",
   "jupyter.about.message": "<Paragraph>Używasz notatnika Jupyter w CoCalc.</Paragraph> <Paragraph>Notatnik Jupyter w CoCalc to całkowicie otwartoźródłowy przepis wykonany przez SageMath, Inc. klasycznego klienta notatnika Jupyter z <A1>projektu Jupyter</A1>. Notatnik Jupyter w CoCalc zachowuje pełną zgodność z formatem pliku oraz ogólnym wyglądem klasycznego notatnika. Ulepsza klasyczny notatnik w następujący sposób: {features} Niektóre funkcje klasycznych rozszerzeń nie są jeszcze wspierane (jeśli czegoś potrzebujesz, <A2>sprawdź tutaj</A2> i utwórz {support_link}), a niektóre z powyższych funkcji są również dostępne w klasycznym Jupyter poprzez rozszerzenia.</Paragraph> <Title>Pytania</Title> {faq} <Title>Informacje o serwerze</Title> {server_info} <Title>Informacje o bieżącym Kernelu</Title> {kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "Wyczyść Wyjście",
   "jupyter.cell-buttonbar-menu.copy-cell": "Kopiuj Komórkę",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "Kopiuj do schowka",
   "jupyter.cell-buttonbar-menu.cut-cell": "Wytnij komórkę",
