+++ src/packages/frontend/i18n/trans/pt_BR.json
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "Turco",
   "jupyter.about.features": "<ul> <li> <b>Sincronização em tempo real:</b> Várias pessoas podem editar notebooks simultaneamente: múltiplos cursores, desfazer e refazer específicos para cada usuário em todo o documento, ipywidgets sincronizados em tempo real </li> <li> <b>Divisão de documento:</b> edite e veja várias partes de um notebook grande ao mesmo tempo </li> <li> <b>Viagem no tempo:</b> veja o histórico detalhado de como exatamente um notebook foi criado </li> <li> <b>Texto/Markdown:</b> editor gráfico poderoso para células de markdown, para que você possa escrever texto bem formatado para explicar seu código sem precisar escrever markdown diretamente. </li> <li> <b>Assistente de IA:</b> Várias ferramentas de IA ajudam a entender mensagens de erro, refinar código, gerar novas células de código com base no contexto ou apenas explicar/documentar o código existente. </li> <li> <b>Trechos:</b> exemplos de código para muitos kernels </li> <li> <b>Quadro branco:</b> crie um quadro branco (documentado com final \".board\") e use células Jupyter como parte do seu quadro branco </li> <li> <b>Zoom:</b> altere facilmente o tamanho da fonte </li> <li> <b>Dobras de código:</b> veja a estrutura da entrada </li> <li> <b>Formatação de código:</b> clique no botão Formatar para formatar automaticamente seu código e markdown. </li> <li> <b>Aparência moderna:</b> botões, menus e dicas de execução de células que refletem melhor o estado </li> <li> <b>Saída grande:</b> limitação, janelamento e buffer no servidor </li> <li> <b>Captura de saída em segundo plano:</b> funciona se nenhum usuário tiver o notebook aberto (<A>discussão</A>) </li> <li> <b>Suporte móvel:</b> telefones e tablets </li> <li> <b>Criação de células:</b> clique na linha azul entre as células para criar novas células </li> <li> <b>Compartilhar:</b> seu trabalho é visível publicamente através do nosso visualizador de notebook rápido e leve </li> <li> <b>LaTeX:</b> exporte o notebook para LaTeX e depois edite o LaTeX gerado diretamente no CoCalc. </li> <li> <b>Atalhos e esquemas de cores:</b> Atalhos do VIM, Emacs e Sublime, e muitos esquemas de cores (nas configurações da conta) </li> </ul>",
   "jupyter.about.message": "<Paragraph>Você está usando o notebook Jupyter do CoCalc.</Paragraph> <Paragraph> O notebook Jupyter do CoCalc é uma reescrita completa de código aberto feita pela SageMath, Inc. do cliente clássico do Jupyter Notebook do <A1>projeto Jupyter</A1>. O notebook Jupyter do CoCalc mantém total compatibilidade com o formato de arquivo e a aparência geral do notebook clássico. Ele melhora o notebook clássico da seguinte forma: {features} Algumas funcionalidades das extensões clássicas ainda não são suportadas (se você precisar de algo, <A2>verifique aqui</A2> e crie um {support_link}), e algumas das funcionalidades acima também estão disponíveis no Jupyter clássico via extensões. </Paragraph> <Title>Perguntas</Title> {faq} <Title>Informações do Servidor</Title> {server_info} <Title>Informações do Kernel Atual</Title> {kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "Limpar Saída",
   "jupyter.cell-buttonbar-menu.copy-cell": "Copiar Célula",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "Copiar para a Área de Transferência",
   "jupyter.cell-buttonbar-menu.cut-cell": "Cortar Célula",
