+++ src/packages/frontend/i18n/trans/pt_PT.json
@@ -597,7 +597,7 @@
   "editor.snippets.label": "Trechos",
   "editor.snippets.tooltip": "Abrir um painel contendo trechos de código",
   "editor.table_of_contents.name": "Índice",
-  "editor.table_of_contents.short": "Conteúdo",
+  "editor.table_of_contents.short": "Conteúdos",
   "editor.terminal.cmd.help.title": "Mostrar documentação para usar o Terminal Linux no CoCalc",
   "file_actions.compress.name": "Comprimir",
   "file_actions.copy.name": "Copiar",
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "Turco",
   "jupyter.about.features": "<ul> <li> <b>Sincronização em tempo real:</b> Várias pessoas podem editar cadernos simultaneamente: múltiplos cursosres, desfazer e refazer específicos de usuário em todo o documento, ipywidgets sincronizados em tempo real </li> <li> <b>Divisão de documento:</b> edite e veja várias partes de um caderno grande ao mesmo tempo </li> <li> <b>TimeTravel:</b> veja o histórico detalhado de como um caderno foi criado </li> <li> <b>Texto/Markdown:</b> editor gráfico poderoso para células markdown, para que possa escrever texto bem formatado para explicar o seu código sem ter que escrever markdown diretamente. </li> <li> <b>Assistente de IA:</b> Várias ferramentas de IA ajudam a entender mensagens de erro, refinar código, gerar novas células de código com base no contexto, ou simplesmente explicar/documentar código existente. </li> <li> <b>Trechos de código:</b> exemplos de código para muitos kernels </li> <li> <b>Quadro branco:</b> crie um quadro branco (documento terminado em \".board\") e use células Jupyter como parte do seu quadro branco </li> <li> <b>Zoom:</b> altere facilmente o tamanho da fonte </li> <li> <b>Dobra de código:</b> veja a estrutura da entrada </li> <li> <b>Formatação de código:</b> clique no botão Formatar para formatar automaticamente o seu código e markdown. </li> <li> <b>Aparência moderna:</b> botões, menus e dicas de execução de células que refletem melhor o estado </li> <li> <b>Saída grande:</b> limitação, janelamento e buffer no lado do servidor </li> <li> <b>Captura de saída em segundo plano:</b> funciona se nenhum usuário tiver o caderno aberto (<A>discussão</A>) </li> <li> <b>Suporte móvel:</b> telemóveis e tablets </li> <li> <b>Criação de células:</b> clique na linha azul entre as células para criar novas células </li> <li> <b>Partilhar:</b> o seu trabalho é visível publicamente através do nosso visualizador de cadernos rápido e leve </li> <li> <b>LaTeX:</b> exporte o caderno para LaTeX, depois edite o LaTeX gerado diretamente no CoCalc. </li> <li> <b>Atalhos de teclado e esquemas de cores:</b> Atalhos de teclado VIM, Emacs e Sublime, e muitos esquemas de cores (nas configurações da conta) </li> </ul>",
   "jupyter.about.message": "<Paragraph>Está a usar o caderno Jupyter do CoCalc</Paragraph> <Paragraph> O caderno Jupyter do CoCalc é uma reescrita completa e de código aberto do cliente Jupyter Notebook clássico, realizada pela SageMath, Inc., do <A1>projeto Jupyter</A1>. O caderno Jupyter do CoCalc mantém total compatibilidade com o formato de ficheiro e a aparência geral do caderno clássico. Melhora o caderno clássico da seguinte forma: {features} Algumas funcionalidades das extensões clássicas ainda não são suportadas (se precisar de algo, <A2>verifique aqui</A2> e crie um {support_link}), e algumas das melhorias acima também estão disponíveis no Jupyter clássico através de extensões. </Paragraph> <Title>Perguntas</Title> {faq} <Title>Informação do Servidor</Title> {server_info} <Title>Informação Atual do Kernel</Title> {kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "Limpar Saída",
   "jupyter.cell-buttonbar-menu.copy-cell": "Copiar Célula",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "Copiar para a Área de Transferência",
   "jupyter.cell-buttonbar-menu.cut-cell": "Cortar Célula",
