+++ src/packages/frontend/i18n/trans/ru_RU.json
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "Турецкий",
   "jupyter.about.features": "<ul> <li> <b>Синхронизация в реальном времени:</b> Несколько человек могут одновременно редактировать блокноты: несколько курсоров, пользовательская отмена и повтор, синхронизация ipywidgets в реальном времени </li> <li> <b>Разделение документа:</b> редактировать и видеть несколько частей большого блокнота одновременно </li> <li> <b>TimeTravel:</b> увидеть подробную историю создания блокнота </li> <li> <b>Текст/Markdown:</b> мощный графический редактор для markdown-ячеек, чтобы вы могли писать хорошо оформленный текст для объяснения кода без необходимости писать markdown напрямую. </li> <li> <b>AI Assistant:</b> Несколько AI-инструментов помогают вам понимать сообщения об ошибках, улучшать код, генерировать новые ячейки кода на основе контекста или просто объяснять/документировать существующий код. </li> <li> <b>Фрагменты:</b> образцы кода для многих ядёр </li> <li> <b>Доска:</b> создайте доску (документ с расширением \".board\") и используйте ячейки Jupyter как часть вашей доски </li> <li> <b>Масштабирование:</b> легко изменить размер шрифта </li> <li> <b>Свертывание кода:</b> видеть структуру ввода </li> <li> <b>Форматирование кода:</b> нажмите кнопку Форматировать, чтобы автоматически форматировать ваш код и markdown. </li> <li> <b>Современный вид:</b> кнопки, меню и подсказки выполнения ячеек, лучше отражающие состояние </li> <li> <b>Большой вывод:</b> серверное дросселирование, оконный режим и буферизация </li> <li> <b>Фоновый захват вывода:</b> работает, если никто не открыл блокнот (<A>обсуждение</A>) </li> <li> <b>Поддержка мобильных устройств:</b> телефоны и планшеты </li> <li> <b>Создание ячеек:</b> нажмите на синюю линию между ячейками, чтобы создать новые ячейки </li> <li> <b>Поделиться:</b> ваша работа видна публично через наш быстрый легковесный просмотрщик блокнотов </li> <li> <b>LaTeX:</b> экспорт блокнота в LaTeX, затем редактируйте созданный LaTeX прямо в CoCalc. </li> <li> <b>Горячие клавиши и цветовые схемы:</b> VIM, Emacs и Sublime горячие клавиши, и множество цветовых схем (в настройках аккаунта) </li> </ul>",
   "jupyter.about.message": "Вы используете блокнот CoCalc Jupyter.\nБлокнот CoCalc Jupyter — это полная открытая переработка SageMath, Inc. классического клиента Jupyter Notebook от проекта <A1>Jupyter</A1>. Блокнот CoCalc Jupyter сохраняет полную совместимость с форматом файлов и общим внешним видом классического блокнота. Он улучшает классический блокнот следующим образом: {features} Некоторая функциональность классических расширений пока не поддерживается (если вам что-то нужно, <A2>проверьте здесь</A2> и создайте {support_link}), и некоторые из вышеупомянутых функций также доступны в классическом Jupyter через расширения.\nВопросы\n{faq}\nИнформация о сервере\n{server_info}\nИнформация о текущем ядре\n{kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "Очистить вывод",
   "jupyter.cell-buttonbar-menu.copy-cell": "Копировать ячейку",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "Копировать в буфер обмена",
   "jupyter.cell-buttonbar-menu.cut-cell": "Вырезать ячейку",
