+++ src/packages/frontend/i18n/trans/tr_TR.json
@@ -597,7 +597,7 @@
   "editor.snippets.label": "Parçalar",
   "editor.snippets.tooltip": "Kod parçacıklarını içeren bir panel aç.",
   "editor.table_of_contents.name": "İçindekiler",
-  "editor.table_of_contents.short": "İçerik",
+  "editor.table_of_contents.short": "İçindekiler",
   "editor.terminal.cmd.help.title": "CoCalc'ta Linux Terminalini kullanma belgelerini göster.",
   "file_actions.compress.name": "Sıkıştır",
   "file_actions.copy.name": "Kopyala",
@@ -652,6 +652,7 @@
   "i18n.localization.lang.turkish": "Türkçe",
   "jupyter.about.features": "<ul> <li> <b>Gerçek zamanlı senkronizasyon:</b> Birden fazla kişi aynı anda defterleri düzenleyebilir: birden fazla imleç, belge genelinde kullanıcıya özel geri al ve yinele, gerçek zamanlı senkronize edilmiş ipywidgets </li> <li> <b>Belge bölme:</b> büyük bir defterin birden fazla kısmını aynı anda düzenleyin ve görün </li> <li> <b>TimeTravel:</b> bir defterin tam olarak nasıl oluşturulduğuna dair ayrıntılı geçmişi görün </li> <li> <b>Metin/Markdown:</b> markdown hücreleri için güçlü grafik düzenleyici, böylece kodunuzu doğrudan markdown yazmak zorunda kalmadan güzel biçimlendirilmiş metinlerle açıklayabilirsiniz </li> <li> <b>AI Asistanı:</b> Birkaç AI aracı hata mesajlarını anlamanıza, kodu geliştirmenize, bağlama dayalı yeni kod hücreleri oluşturmanıza veya mevcut kodu sadece açıklamanıza/belgelendirmenize yardımcı olur </li> <li> <b>Snippets:</b> birçok çekirdek için kod örnekleri </li> <li> <b>Beyaz tahta:</b> bir beyaz tahta oluşturun (\".board\" ile biten belgeler) ve beyaz tahtanızın bir parçası olarak Jupyter hücrelerini kullanın </li> <li> <b>Yakınlaştırma:</b> yazı tipi boyutunu kolayca değiştirin </li> <li> <b>Kod katlama:</b> girdi yapısını görün </li> <li> <b>Kod biçimlendirme:</b> kodunuzu ve markdown'ı otomatik olarak biçimlendirmek için Biçimlendir düğmesine tıklayın </li> <li> <b>Modern görünüm:</b> durumu daha iyi yansıtan düğmeler, menüler ve hücre yürütme ipuçları </li> <li> <b>Büyük çıktı:</b> sunucu tarafı kısıtlaması, pencereler ve tamponlama </li> <li> <b>Arka planda çıktı yakalama:</b> defter açık değilse çalışır (<A>tartışma</A>) </li> <li> <b>Mobil destek:</b> telefonlar ve tabletler </li> <li> <b>Hücre oluşturma:</b> yeni hücreler oluşturmak için hücreler arasındaki mavi çizgiye tıklayın </li> <li> <b>Paylaş:</b> çalışmanız hızlı ve hafif defter görüntüleyicimiz aracılığıyla genel olarak görünebilir </li> <li> <b>LaTeX:</b> defteri LaTeX'e aktarın, ardından oluşturulan LaTeX'i doğrudan CoCalc'ta düzenleyin </li> <li> <b>Klavye kısayolları ve renk şemaları:</b> VIM, Emacs ve Sublime klavye kısayolları ve birçok renk şeması (hesap ayarlarında) </li> </ul>",
   "jupyter.about.message": "<Paragraph>CoCalc Jupyter defterini kullanıyorsunuz.</Paragraph> <Paragraph>CoCalc Jupyter Defteri, SageMath, Inc. tarafından Jupyter projesinden klasik Jupyter Defter istemcisinin tamamen açık kaynaklı bir yeniden yazımıdır. CoCalc Jupyter Defteri, dosya formatı ve genel görünüm ile klasik defterle tam uyumluluğunu korur. Klasik deftere göre şu şekilde iyileştirilmiştir: {features} Klasik uzantıların bazı işlevsellikleri henüz desteklenmiyor (bir şeye ihtiyacınız varsa, <A2>burayı kontrol edin</A2> ve bir {support_link} oluşturun) ve yukarıdakilerin bazıları klasik Jupyter'de uzantılar aracılığıyla da mevcuttur.</Paragraph> <Title>Sorular</Title> {faq} <Title>Sunucu Bilgileri</Title> {server_info} <Title>Mevcut Çekirdek Bilgileri</Title> {kernel_info}",
+  "jupyter.cell-buttonbar-menu.clear-output": "Çıktıyı Temizle",
   "jupyter.cell-buttonbar-menu.copy-cell": "Hücresi Kopyala",
   "jupyter.cell-buttonbar-menu.copy-clipboard": "Panoya Kopyala",
   "jupyter.cell-buttonbar-menu.cut-cell": "Hücreyi Kes",
